Revert "Remove last dep of spack.environment on spack.cmd"
This reverts commit 89177317cc2b50e2845bf968507c8afa3a0ab0a5.
